removed support for package parameters and package script-based wf config generation
